refactor: Move the server logic into its own module
